qbo_checker/requirements.txt
playwright>=1.40.0
Pillow>=10.0.0
openpyxl>=3.1.0
anthropic>=0.18.0
